Added TODO and integrated maze generator to main.
Added TODO's for some maze generator problems, to be addressed at a later date. Maze generator now is integrated with the search algorithm, but can also be run as a script. Changed some of the functionality around what types of objects are being passed along. Squares traversed are now represented in PPM with the color yellow.